Gradle Support
References #31

You need to configure the project as 'gradle nature' on your IDE, to
allow it find the referenced libraries


use 'gradlew' on Unix system, and gradlew.bat on Windows

To build a dist
./gradlew dist 

To compile only L2jBrasil.jar
./gradlew jar


#References #24

Changed logback api to log4j, so the HikariCP will not send logs to
console. However, we need to configure the log4j and probably a full log
system rework.